Move Types out of public API.
We may add it back later at some point, but I'd like to give the design
more thought before doing that.
